Add files via upload
This is a directory which contains the code crafted for the Algorithm I course, built by Kevin Wayne and Robert Sedgerwick. I've uploaded
only the code of data structures and the algorithms, later i'll update the code for the exercises presents in the course (due to the closeness
of the Mooc, at the date i'm uploading).